+++ src/Public/Add-DataFlow.ps1
@@ -19,42 +19,53 @@ function Add-DataFlow {
         $Journey,
 
         # The identifier key of the data flow.
-        [Parameter(Mandatory, Position = 0)]
+        [Parameter(Mandatory, Position = 0, ParameterSetName = 'Properties')]
         [ValidateNotNullOrEmpty()]
         [ValidateScript({ $_ -notmatch ' ' }, ErrorMessage = 'Value must not contain spaces.')]
         [string] $Key,
 
         # The title of the data flow.
-        [Parameter(Position = 1)]
+        [Parameter(Position = 1, ParameterSetName = 'Properties')]
         [ValidateNotNullOrEmpty()]
         [string] $Title,
 
         # The source models of tha data flow
-        [Parameter(ValueFromPipeline, Mandatory)]
+        [Parameter(ValueFromPipeline, Mandatory, ParameterSetName = 'Properties')]
         [ValidateNotNullOrEmpty()]
-        [string[]] $Source,
+        [Alias('Source')]
+        [string[]] $Sources,
 
         # The sink models of tha data flow
-        [Parameter(ValueFromPipeline, Mandatory)]
+        [Parameter(ValueFromPipeline, Mandatory, ParameterSetName = 'Properties')]
         [ValidateNotNullOrEmpty()]
-        [string[]] $Sink
+        [Alias('Sink')]
+        [string[]] $Sinks,
+
+        # The data flow is added to the data journey.
+        [Parameter( Mandatory, ParameterSetName = 'InputObject' )]
+        [PSCustomObject] $InputObject,
+
+        # Switch that specifies, if the model should be returned instead of only added to the data journey.
+        [Parameter()]
+        [switch] $PassThru
     )
 
     process {
-        $flow = [PSCustomObject]@{
-            Key = $Key
-            Sources = $Source
-            Sinks = $Sink
+        switch ($PsCmdlet.ParameterSetName) {
+            Properties {
+                $InputObject = New-DataFlow -Key:$Key -Title:$Title -Sources:$Sources -Sinks:$Sinks
+            }
+            InputObject {
+            }
+            default {
+                Write-Error "ParameterSetName '$_' not supported"
+            }
         }
 
-        if ( $Title ) {
-            $flow | Add-Member Title $Title
-        }
-
-        $Journey.Flows += $flow
+        $Journey.Flows += $InputObject
 
         if ( $PassThru.IsPresent ) {
-            Write-Output $flow
+            Write-Output $InputObject
         }
     }
 }